refactor: api 호출 함수 한 파일에 한 export로 변경
