+++ frontend/techpick/src/apis/pick/getPickRecordByFolderId.ts
@@ -1,0 +1,7 @@
+import { generatePickRecordData } from '@/utils/generatePickRecordData';
+import { getPickListByFolderId } from './getPickListByFolderId';
+
+export const getPickRecordByFolderId = async (folderId: number) => {
+  const data = await getPickListByFolderId(folderId);
+  return generatePickRecordData(data[0].pickList);
+};
